requirements.txt
python-vlc==3.0.11115
pyusb==1.1.1
